Add new command line feature called kernelreport which generates an email ready report for a set of Android Common Kernels from qa-reports.linaro.org.
ex:
lkftreport.sh kernelreport 4.4 2> /dev/null

This will for all 4.4 combinations, find the 2 most recent "good" runs
which is defined as 2 fully complete jobs with results.
Next it will compute the regressions between the two jobs and output
to stdout a snipping that documents kernel, os, board, kernel versions
compared and the list of regressions.

Note that I recommend piping stderr to null for the time being as the
tool is a bit chatty.