This modifies the name of individual test cases to make sure they either have _32bit, _64bit or the abi appended at the end of the name. This gives us test names of the form
Module#test#abi

This also fixes a bug where not all failures were being found. A
failure that was affecting both 32 bit and 64 bit would be combined
into one if they had the same name. While this is sometimes a good
behavior there is a risk that a 32 bit failure and a 64 bit failure
are not the same.

Signed-off-by: Tom Gall <tom.gall@linaro.org>